+++ classes/suggested-tasks/local-tasks/providers/class-search-engine-visibility.php
@@ -1,0 +1,63 @@
+<?php
+/**
+ * Add tasks for settings saved.
+ *
+ * @package Progress_Planner
+ */
+
+namespace Progress_Planner\Suggested_Tasks\Local_Tasks\Providers;
+
+/**
+ * Add tasks to check if WP debug is enabled.
+ */
+class Search_Engine_Visibility extends Local_OneTime_Tasks_Abstract {
+
+	/**
+	 * The provider type.
+	 *
+	 * @var string
+	 */
+	const TYPE = 'configuration';
+
+	/**
+	 * The provider ID.
+	 *
+	 * @var string
+	 */
+	const ID = 'search-engine-visibility';
+
+	/**
+	 * Check if the task should be added.
+	 *
+	 * @return bool
+	 */
+	public function should_add_task() {
+		return 0 === (int) \get_option( 'blog_public' );
+	}
+
+	/**
+	 * Get the task details.
+	 *
+	 * @param string $task_id The task ID.
+	 *
+	 * @return array
+	 */
+	public function get_task_details( $task_id = '' ) {
+
+		if ( ! $task_id ) {
+			$task_id = $this->get_provider_id();
+		}
+
+		return [
+			'task_id'     => $task_id,
+			'title'       => \esc_html__( 'Allow your site to be indexed by search engines', 'progress-planner' ),
+			'parent'      => 0,
+			'priority'    => 'high',
+			'type'        => $this->get_provider_type(),
+			'points'      => 1,
+			'url'         => $this->capability_required() ? \esc_url( \admin_url( 'options-reading.php' ) ) : '',
+			'dismissible' => true,
+			'description' => '<p>' . \esc_html__( 'Your site is not currently visible to search engines. Consider allowing search engines to index your site.', 'progress-planner' ) . '</p>',
+		];
+	}
+}